fix bug

+++ models/clustering.py
@@ -46,7 +46,7 @@ def compute_features(dataloader, cluster_network, opt):
     iterator.set_description('Computing features & Clustering... ')
 
     num_samples = len(dataloader)
-    feature_coll = np.zeros((num_samples * 49, opt.feature_dim))
+    feature_coll = np.zeros((num_samples * 49, opt.feature_dim), dtype='float32')
     last_idx = 0
     for i, input_image in enumerate(iterator):
         input_image = input_image.to(opt.device)
@@ -92,7 +92,7 @@ def cluster(features, num_cluster):
     nsample, d = features.shape
     niter = int(20 * (nsample / (1000 * num_cluster))) # As a rule of thumb there is no consistent improvement of the k-means quantizer beyond 20 iterations and 1000 * k training points
     # faiss implementation of k-means
-    clus = faiss.Kmeans(d, num_cluster, gpu=True, spherical=True, seed=42, niter = niter)
+    clus = faiss.Kmeans(d, num_cluster, gpu=True, spherical=True, seed=42, niter=niter)
 
     # Change faiss seed at each k-means so that the randomly picked
     # initialization centroids do not correspond to the same feature ids
